pyproject.toml
[project]
name = "upwork-jobs"
version = "0.2.0"
description = "Upwork job scraper - fetches and stores job postings"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.9.0",
    "psycopg[binary,pool]>=3.2",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=24.10.0",
    "isort>=5.13.2",
    "pylint>=3.3.3",
    "pytest>=8.0",
    "pytest-xdist>=3.0",
]
